+++ Projet perso/Node.cpp
@@ -2,7 +2,6 @@
 
 #include <algorithm>
 #include <cassert>
-#include <cstdint>
 #include <functional>
 
 
@@ -11,20 +10,20 @@
 class Counter {
 private:
 	std::vector<bool> loopGest;
-	uint64_t size;
-	std::function<bool(const std::vector<bool>&, uint64_t)> func;
+	size_t size;
+	std::function<bool(const std::vector<bool>&, size_t)> func;
 
 public:
-	Counter(uint64_t size, const std::function<bool(const std::vector<bool>&, uint64_t)>& func);
+	Counter(size_t size, const std::function<bool(const std::vector<bool>&, size_t)>& func);
 	~Counter();
 	std::vector<bool> getNext();
 	bool finished() const;
 	std::vector<std::vector<bool>> generateAll();
 
 private:
-	uint64_t state();
+	size_t state();
 	void reset();
-	void __next(uint64_t i);
+	void __next(size_t i);
 	void findNext();
 	void next();
 
@@ -32,10 +31,10 @@ private:
 
 class BinomialConfigs : public Counter {
 public:
-	BinomialConfigs(uint64_t n, uint64_t m) :
-		Counter(m, [n, m](const std::vector<bool>& v, uint64_t s) {
-			        uint64_t count = 0;
-			        for (uint64_t i = 0; i < s; i++)
+	BinomialConfigs(size_t n, size_t m) :
+		Counter(m, [n, m](const std::vector<bool>& v, size_t s) {
+			        size_t count = 0;
+			        for (size_t i = 0; i < s; i++)
 				        if (v[i]) count++;
 
 			        return count == n;
@@ -46,14 +45,14 @@ public:
 
 class AllConfigs : public Counter {
 public:
-	AllConfigs(uint64_t s) :
-		Counter(s, [](const std::vector<bool>&, uint64_t) {
+	AllConfigs(size_t s) :
+		Counter(s, [](const std::vector<bool>&, size_t) {
 			        return true;
 		        }) {
 	};
 };
 
-Counter::Counter(uint64_t size, const std::function<bool(const std::vector<bool>&, uint64_t)>& func) : size(size),
+Counter::Counter(size_t size, const std::function<bool(const std::vector<bool>&, size_t)>& func) : size(size),
                                                                                                        func(func),
                                                                                                        loopGest(size + 1, false) {
 
@@ -62,9 +61,9 @@ Counter::Counter(uint64_t size, const std::function<bool(const std::vector<bool>
 Counter::~Counter() {
 }
 
-uint64_t Counter::state() {
-	uint64_t res = 0;
-	for (int64_t i = loopGest.size() - 1; i >= 0; i--) {
+size_t Counter::state() {
+	size_t res = 0;
+	for (int i = loopGest.size() - 1; i >= 0; i--) {
 		res *= 2;
 		if (loopGest[i]) {
 			res++;
@@ -77,7 +76,7 @@ uint64_t Counter::state() {
 std::vector<bool> Counter::getNext() {
 	std::cout << "\tConfiguration Found: " << state() << std::endl;
 	std::vector<bool> res(size);
-	for (uint64_t i = 0; i < size; i++) {
+	for (size_t i = 0; i < size; i++) {
 		res[i] = loopGest[i];
 	}
 	next();
@@ -97,13 +96,13 @@ void Counter::next() {
 
 
 void Counter::reset() {
-	for (uint64_t i = 0; i < size + 1; i++) {
+	for (size_t i = 0; i < size + 1; i++) {
 		loopGest[i] = false;
 	}
 	findNext();
 }
 
-void Counter::__next(uint64_t i) {
+void Counter::__next(size_t i) {
 	if (i < loopGest.size()) {
 		if (loopGest[i]) {
 			loopGest[i] = false;
@@ -164,9 +163,9 @@ std::vector<OrientationOnNode> Node::allPossibleOrientations() const {
 	std::vector<Edge> notLockedLoopedEdges;
 	std::vector<Edge> notLockedNotLoopedEdges;
 
-	uint64_t degree = 0;
-	uint64_t in_degree = 0;
-	uint64_t out_degree = 0;
+	size_t degree = 0;
+	size_t in_degree = 0;
+	size_t out_degree = 0;
 
 	std::for_each(edges.begin(), edges.end(), [&degree](const Edge* ed) {
 		              if (!ed->isLoop()) degree++;
@@ -203,16 +202,16 @@ std::vector<OrientationOnNode> Node::allPossibleOrientations() const {
 	assert(degree % 2 == 0);
 	assert(degree >= out_degree + in_degree);
 
-	// uint64_t is of degree 2.
-	uint64_t demi_degree = degree / 2;
+	// size_t is of degree 2.
+	size_t demi_degree = degree / 2;
 
 	if ((in_degree > demi_degree) || (out_degree > demi_degree)) {
 		// No valid orientation
 		assert(orientations.empty());
 		return orientations;
 	} else {
-		uint64_t in_rem = demi_degree - in_degree;
-		uint64_t out_rem = demi_degree - out_degree;
+		size_t in_rem = demi_degree - in_degree;
+		size_t out_rem = demi_degree - out_degree;
 
 		assert(in_rem + out_rem == notLockedNotLoopedEdges.size());
 
@@ -226,7 +225,7 @@ std::vector<OrientationOnNode> Node::allPossibleOrientations() const {
 		//--- Lambdas ---
 		auto applyInChange = [&](std::vector<bool>& setOri, std::vector<Edge>& edges) {
 			assert(setOri.size() == edges.size());
-			for (uint64_t i = 0; i < setOri.size(); i++) {
+			for (size_t i = 0; i < setOri.size(); i++) {
 				assert(edges[i].isAdjacentTo(this));
 				if (setOri[i]) {
 					if (edges[i].getDestination() != this) {
@@ -241,7 +240,7 @@ std::vector<OrientationOnNode> Node::allPossibleOrientations() const {
 		};
 		auto applyChange = [&](std::vector<bool>& setOri, std::vector<Edge>& edges) {
 			assert(setOri.size() == edges.size());
-			for (uint64_t i = 0; i < setOri.size(); i++) {
+			for (size_t i = 0; i < setOri.size(); i++) {
 				if (setOri[i]) {
 					edges[i].reverseOrientation();
 				}
@@ -266,7 +265,7 @@ std::vector<OrientationOnNode> Node::allPossibleOrientations() const {
 		if (loopedStates.empty()) {
 			// There is no loop
 			std::cout << "No loop found..." << std::endl;
-			for (uint64_t i = 0; i < notLoopedStates.size(); i++) {
+			for (size_t i = 0; i < notLoopedStates.size(); i++) {
 				resEdges.clear();
 				std::vector<Edge> othersEdges(notLockedNotLoopedEdges);
 				assert(notLoopedStates[i].size() == othersEdges.size());
@@ -280,7 +279,7 @@ std::vector<OrientationOnNode> Node::allPossibleOrientations() const {
 		} else if (notLoopedStates.empty()) {
 			// There is nothing else but loop
 			std::cout << "Only loops found..." << std::endl;
-			for (uint64_t j = 0; j < loopedStates.size(); j++) {
+			for (size_t j = 0; j < loopedStates.size(); j++) {
 				resEdges.clear();
 				std::vector<Edge> loops(notLockedLoopedEdges);
 				assert(loopedStates[j].size() == loops.size());
@@ -293,8 +292,8 @@ std::vector<OrientationOnNode> Node::allPossibleOrientations() const {
 			}
 		} else {
 			std::cout << "Loops and other edges found..." << std::endl;
-			for (uint64_t i = 0; i < notLoopedStates.size(); i++) {
-				for (uint64_t j = 0; j < loopedStates.size(); j++) {
+			for (size_t i = 0; i < notLoopedStates.size(); i++) {
+				for (size_t j = 0; j < loopedStates.size(); j++) {
 					resEdges.clear();
 					std::vector<Edge> loops(notLockedLoopedEdges);
 					std::vector<Edge> othersEdges(notLockedNotLoopedEdges);
@@ -329,7 +328,7 @@ void Node::setOrientedEdges(const OrientationOnNode& ori) {
 	assert(std::is_sorted(ori.begin(),ori.end()));
 	assert(ori.size() == edges.size());
 
-	for (uint64_t i = 0; i < edges.size(); i++) {
+	for (size_t i = 0; i < edges.size(); i++) {
 		assert(edges[i]->id == ori[i].id);
 		*(edges[i]) = ori[i];
 	}
